requirements.txt
Flask
flask-cors
requests
pinecone-client
